requirements.txt
asgiref==2.3.2
async-timeout==3.0.0
asyncio==3.4.3
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
constantly==15.1.0
daphne==2.2.1
Django==2.1
django-filter==2.0.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
Markdown==2.6.11
mongoengine==0.15.3
numpy==1.15.1
psycopg2==2.7.5
PyHamcrest==1.9.0
pymongo==3.7.1
pytz==2018.5
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
Twisted==18.7.0
txaio==18.7.1
zope.interface==4.5.0
